group internal root tests into subtests
