+++ writer_encoding_internal_test.go
